Implemented backend functionality for the store feature.

+++ backend/src/main/java/com/leandroadal/vortasks/dto/social/GroupTaskDTO.java
@@ -10,7 +10,7 @@ public record GroupTaskDTO(
         String name,
         String description,
         int xp,
-        float coins,
+        int coins,
         String type,
         int repetition,
         String reminder,
