+++ backend/src/main/java/com/leandroadal/vortasks/dto/social/OnlineMissionDTO.java
@@ -10,7 +10,7 @@ public record OnlineMissionDTO(
         String title,
         String description,
         int xp,
-        float coins,
+        int coins,
         String type,
         String repetition,
         String reminder,
